Split monthFuncs into separate functions in month_funcs directory
